Add --test and --cgviz flags
If --test is given, data is added to "test" species in cgviz and no data
is uploaded to CDM

The --cgviz flag is required to upload data to cgviz. Added it as default
to the hopper config file.

I had to decouple the CDM and cgviz processes. So the run folder regex is
currently duplicated. Should be fixed at some point.